Add sortings images and tags: newer files are displaed first
