+++ src/MemeStorage.java
@@ -15,6 +15,8 @@ import java.io.*;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Comparator;
 import java.util.Random;
 
 public class MemeStorage extends JFrame {
@@ -22,7 +24,7 @@ public class MemeStorage extends JFrame {
     JPanel mainPanel = null;
     JScrollPane scrollPane = null;
     String defaultImagesFormat = "png";
-    final String VERSION = "0.3.5";
+    final String VERSION = "0.4";
 
     public MemeStorage() {
         setTitle(programName);
@@ -113,11 +115,7 @@ public class MemeStorage extends JFrame {
         });
         propertiesMenu.add(info);
 
-        try {
-            checkFiles();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        checkFiles();
 
         showAllImages();
 
@@ -219,88 +217,15 @@ public class MemeStorage extends JFrame {
         mainPanel.removeAll();
 
         JPanel contentPanel = new JPanel(new GridLayout(0,5));
-        File[] filesArr = new File("storage/images").listFiles();
+        File[] files = new File("storage/images").listFiles();
 
-        for(int i = 0; i < filesArr.length; i++)
+        Arrays.sort(files, (f1, f2) -> Long.valueOf(f2.lastModified()).compareTo(f1.lastModified()));
+
+        for(int i = 0; i < files.length; i++)
         {
             try {
-                BufferedImage img = ImageIO.read(filesArr[i]);
-                img = scale(img, 90);
-                ImageIcon icon = new ImageIcon(img);
-
-                JLabel imageLabel = new JLabel();
-                int finalI = i;
-                File imageFile = filesArr[i];
-                imageLabel.addMouseListener(new MouseListener() {
-                    @Override
-                    public void mouseClicked(MouseEvent mouseEvent) {
-                        if(mouseEvent.getButton() == MouseEvent.BUTTON1){
-                            Toolkit.getDefaultToolkit().getSystemClipboard().setContents(
-                                    new ImageTransferable(imageFile), null);
-                        }
-                        if(mouseEvent.getButton() == MouseEvent.BUTTON3){
-                            JPopupMenu imageSettingsMenu = new JPopupMenu();
-
-                            JMenuItem deleteImage = new JMenuItem("Delete");
-                            deleteImage.addActionListener(new ActionListener() {
-                                @Override
-                                public void actionPerformed(ActionEvent actionEvent) {
-                                    imageFile.delete();
-                                    JOptionPane.showMessageDialog(imageLabel, "Image successful delete", "Delete image", JOptionPane.INFORMATION_MESSAGE);
-                                    showAllImages();
-                                }
-                            });
-                            imageSettingsMenu.add(deleteImage);
-
-                            JMenuItem tagsImage = new JMenuItem("Tags");
-                            tagsImage.addActionListener(new ActionListener() {
-                                @Override
-                                public void actionPerformed(ActionEvent actionEvent) {
-                                    TagsFrame tagsFrame = new TagsFrame(imageFile);
-                                    tagsFrame.setBounds(imageLabel.getBounds().x, imageLabel.getBounds().y, 400, 400);
-                                }
-                            });
-                            imageSettingsMenu.add(tagsImage);
-
-                            JMenuItem infoImage = new JMenuItem("Info");
-                            infoImage.addActionListener(new ActionListener() {
-                                @Override
-                                public void actionPerformed(ActionEvent actionEvent) {
-                                    JOptionPane.showMessageDialog(imageLabel,  "Name " + imageFile.getName() + "\n" +
-                                                                                        "Path " + imageFile.getAbsolutePath() + "\n" +
-                                                                                        "Size " + (float)(imageFile.length()/1024) + " KB");
-                                }
-                            });
-                            imageSettingsMenu.add(infoImage);
-
-                            imageSettingsMenu.show(imageLabel,0, 50);
-                        }
-                    }
-
-                    @Override
-                    public void mousePressed(MouseEvent mouseEvent) {
-
-                    }
-
-                    @Override
-                    public void mouseReleased(MouseEvent mouseEvent) {
-
-                    }
-
-                    @Override
-                    public void mouseEntered(MouseEvent mouseEvent) {
-
-                    }
-
-                    @Override
-                    public void mouseExited(MouseEvent mouseEvent) {
-
-                    }
-                });
-                imageLabel.setIcon(icon);
-                contentPanel.add(imageLabel);
+                ImageLabel imageLabel = new ImageLabel(files[i], contentPanel);
                 mainPanel.add(contentPanel);
-
             } catch (IOException e) {
                 e.printStackTrace();
             }
@@ -315,7 +240,14 @@ public class MemeStorage extends JFrame {
         mainPanel.setLayout(new BorderLayout());
 
         JPanel contentPanel = new JPanel(new GridLayout(0,5));
-        String[] tagsFile = new File("storage/tags/").list();
+        File[] filesArr = new File("storage/tags/").listFiles();
+
+        Arrays.sort(filesArr, (f1, f2) -> Long.valueOf(f2.lastModified()).compareTo(f1.lastModified()));
+
+        String[] tagsFile = new String[filesArr.length];
+        for(int i = 0; i < tagsFile.length; i++)
+            tagsFile[i] = filesArr[i].getName();
+
         if(tagsFile.length > 0){
             JList<String> tagsList = new JList<>(tagsFile);
 
@@ -332,95 +264,37 @@ public class MemeStorage extends JFrame {
             mainPanel.add(tagsScrollPane, BorderLayout.WEST);
             mainPanel.add(contentScrollPane, BorderLayout.CENTER);
 
-            tagsList.addListSelectionListener(new ListSelectionListener() {
-                @Override
-                public void valueChanged(ListSelectionEvent listSelectionEvent) {
-                    try {
-                        contentPanel.removeAll();
-                        contentPanel.revalidate();
+            tagsList.addListSelectionListener(listSelectionEvent -> {
+                try {
+                    contentPanel.removeAll();
+                    contentPanel.revalidate();
 
-                        BufferedReader reader = new BufferedReader(new FileReader("storage/tags/" + tagsList.getSelectedValue()));
+                    BufferedReader reader = new BufferedReader(new FileReader("storage/tags/" + tagsList.getSelectedValue()));
 
-                        String line;
-                        while ((line = reader.readLine()) != null) {
-                            File image = new File("storage/images/" + line);
+                    ArrayList<File> tagedImage = new ArrayList<>();
 
-                            BufferedImage img = ImageIO.read(image);
-                            img = scale(img, 90);
-                            ImageIcon icon = new ImageIcon(img);
+                    String line;
+                    while ((line = reader.readLine()) != null) {
+                        File image = new File("storage/images/" + line);
+                        tagedImage.add(image);
+                    }
 
-                            JLabel imageLabel = new JLabel();
-                            imageLabel.addMouseListener(new MouseListener() {
-                                @Override
-                                public void mouseClicked(MouseEvent mouseEvent) {
-                                    if(mouseEvent.getButton() == MouseEvent.BUTTON1){
-                                        Toolkit.getDefaultToolkit().getSystemClipboard().setContents(
-                                                new ImageTransferable(image), null);
-                                    }
-                                    if(mouseEvent.getButton() == MouseEvent.BUTTON3){
-                                        JPopupMenu imageSettingsMenu = new JPopupMenu();
+                    tagedImage.sort((f1, f2) -> Long.valueOf(f2.lastModified()).compareTo(f1.lastModified()));
 
-                                        JMenuItem deleteImage = new JMenuItem("Delete");
-                                        deleteImage.addActionListener(new ActionListener() {
-                                            @Override
-                                            public void actionPerformed(ActionEvent actionEvent) {
-                                                image.delete();
-                                                JOptionPane.showMessageDialog(imageLabel, "Image successful delete", "Delete image", JOptionPane.INFORMATION_MESSAGE);
-                                                showAllImages();
-                                            }
-                                        });
-                                        imageSettingsMenu.add(deleteImage);
+                    for(int i = 0; i < tagedImage.size(); i++) {
+                        File image = tagedImage.get(i);
+                        ImageLabel imageLabel = new ImageLabel(image, contentPanel);
+                    }
 
-                                        JMenuItem tagsImage = new JMenuItem("Tags");
-                                        tagsImage.addActionListener(new ActionListener() {
-                                            @Override
-                                            public void actionPerformed(ActionEvent actionEvent) {
-                                                TagsFrame tagsFrame = new TagsFrame(image);
-                                                tagsFrame.setBounds(imageLabel.getBounds().x, imageLabel.getBounds().y, 400, 400);
-                                            }
-                                        });
-                                        imageSettingsMenu.add(tagsImage);
-
-                                        JMenuItem infoImage = new JMenuItem("Info");
-                                        infoImage.addActionListener(new ActionListener() {
-                                            @Override
-                                            public void actionPerformed(ActionEvent actionEvent) {
-                                                JOptionPane.showMessageDialog(imageLabel,  "Name " + image.getName() + "\n" +
-                                                        "Path " + image.getAbsolutePath() + "\n" +
-                                                        "Size " + (float)(image.length()/1024) + " KB");
-                                            }
-                                        });
-                                        imageSettingsMenu.add(infoImage);
-
-                                        imageSettingsMenu.show(imageLabel,0, 50);
-                                    }
-                                }
-
-                                @Override
-                                public void mousePressed(MouseEvent mouseEvent) {}
-
-                                @Override
-                                public void mouseReleased(MouseEvent mouseEvent) {}
-
-                                @Override
-                                public void mouseEntered(MouseEvent mouseEvent) {}
-
-                                @Override
-                                public void mouseExited(MouseEvent mouseEvent) {}
-                            });
-                            imageLabel.setIcon(icon);
-                            contentPanel.add(imageLabel);
-                        }
-                        contentPanel.revalidate();
-                        contentPanel.repaint();
-                    } catch (FileNotFoundException e) {
-                        e.printStackTrace();
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                        int confirmDelete = JOptionPane.showConfirmDialog(null, "This tag is empty or damaged. Delete this?", "Tag is empty", JOptionPane.YES_NO_OPTION);
-                        if(confirmDelete == JOptionPane.YES_OPTION){
-                            new File("storage/tags/" + tagsList.getSelectedValue()).delete();
-                        }
+                    contentPanel.revalidate();
+                    contentPanel.repaint();
+                } catch (FileNotFoundException e) {
+                    e.printStackTrace();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                    int confirmDelete = JOptionPane.showConfirmDialog(null, "This tag is empty or damaged. Delete this?", "Tag is empty", JOptionPane.YES_NO_OPTION);
+                    if(confirmDelete == JOptionPane.YES_OPTION){
+                        new File("storage/tags/" + tagsList.getSelectedValue()).delete();
                     }
                 }
             });
@@ -449,82 +323,7 @@ public class MemeStorage extends JFrame {
         return name;
     }
 
-    BufferedImage scale(BufferedImage img, int minSize) {
-        int targetWidth = 0;
-        int targetHeight = 0;
-
-        if(img.getHeight() <= minSize && img.getWidth() <= minSize){
-            return img;
-        }else{
-            if(img.getWidth() > img.getHeight()){
-                float scaleKoef = img.getWidth()/minSize;
-
-                targetWidth = (int) (img.getWidth()/scaleKoef);
-                targetHeight = (int) (img.getHeight()/scaleKoef);
-            }else{
-                float scaleKoef = img.getHeight()/minSize;
-
-                targetWidth = (int) (img.getWidth()/scaleKoef);
-                targetHeight = (int) (img.getHeight()/scaleKoef);
-            }
-
-        }
-
-
-        int type = (img.getTransparency() == Transparency.OPAQUE) ? BufferedImage.TYPE_INT_RGB : BufferedImage.TYPE_INT_ARGB;
-        BufferedImage ret = img;
-        BufferedImage scratchImage = null;
-        Graphics2D g2 = null;
-
-        int w = img.getWidth();
-        int h = img.getHeight();
-
-        int prevW = w;
-        int prevH = h;
-
-        do {
-            if (w > targetWidth) {
-                w /= 2;
-                w = (w < targetWidth) ? targetWidth : w;
-            }
-
-            if (h > targetHeight) {
-                h /= 2;
-                h = (h < targetHeight) ? targetHeight : h;
-            }
-
-            if (scratchImage == null) {
-                scratchImage = new BufferedImage(w, h, type);
-                g2 = scratchImage.createGraphics();
-            }
-
-            g2.setRenderingHint(RenderingHints.KEY_INTERPOLATION,
-                    RenderingHints.VALUE_INTERPOLATION_BILINEAR);
-            g2.drawImage(ret, 0, 0, w, h, 0, 0, prevW, prevH, null);
-
-            prevW = w;
-            prevH = h;
-            ret = scratchImage;
-        } while (w != targetWidth || h != targetHeight);
-
-        if (g2 != null) {
-            g2.dispose();
-        }
-
-        if (targetWidth != ret.getWidth() || targetHeight != ret.getHeight()) {
-            scratchImage = new BufferedImage(targetWidth, targetHeight, type);
-            g2 = scratchImage.createGraphics();
-            g2.drawImage(ret, 0, 0, null);
-            g2.dispose();
-            ret = scratchImage;
-        }
-
-
-        return ret;
-
-    }
-
-    void checkFiles() throws IOException {
+    void checkFiles() {
         if(!new File("storage/").exists())
             new File("storage/").mkdir();
